+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Enabled fine-tuning of encoding behavior through `self.crf`, `self.movflags`, and `self.max_muxing_queue_size`.
 
+### Changed
+
+- Allowed custom thumbnail extraction by letting thumb accept a timestamp string ("HH:MM:SS") instead of only True or False.
+
 ## [1.0.0] - 2025-06-04
 
 ### Added